✨ Add BackgroundBeams component for animated SVG beams as a background.
