+++ app/_components/animations/backgrounds/index.js
@@ -1,0 +1,2 @@
+export { default as BackgroundBeams } from "./background-beams";
+export { default as Particles } from "./particles";
