+++ app/_components/animations/text/flip-words.js
@@ -1,0 +1,86 @@
+"use client";
+import React, { useCallback, useEffect, useRef, useState } from "react";
+import { AnimatePresence, motion, LayoutGroup } from "framer-motion";
+import { cn } from "@/utils/class-name";
+
+/**
+ * Renders a component that flips between words with a specified duration.
+ *
+ * @param {Object} props - The properties object.
+ * @param {Array<string>} props.words - The array of words to flip between.
+ * @param {number} [props.duration=3000] - The duration of each word flip in milliseconds.
+ * @param {string} [props.className] - The CSS class name for the component.
+ * @return {JSX.Element} The rendered component.
+ * @example
+ * <FlipWords words={["Word 1", "Word 2", "Word 3"]} />
+ */
+export const FlipWords = ({ words, duration = 3000, className }) => {
+	const [currentWord, setCurrentWord] = useState(words[0]);
+	const [isAnimating, setIsAnimating] = useState(false);
+
+	const startAnimation = useCallback(() => {
+		const word = words[words.indexOf(currentWord) + 1] || words[0];
+		setCurrentWord(word);
+		setIsAnimating(true);
+	}, [currentWord, words]);
+
+	useEffect(() => {
+		if (!isAnimating)
+			setTimeout(() => {
+				startAnimation();
+			}, duration);
+	}, [isAnimating, duration, startAnimation]);
+
+	return (
+		<AnimatePresence
+			onExitComplete={() => {
+				setIsAnimating(false);
+			}}
+		>
+			<motion.div
+				initial={{
+					opacity: 0,
+					y: 10,
+				}}
+				animate={{
+					opacity: 1,
+					y: 0,
+				}}
+				transition={{
+					type: "spring",
+					stiffness: 100,
+					damping: 10,
+				}}
+				exit={{
+					opacity: 0,
+					y: -40,
+					x: 40,
+					filter: "blur(8px)",
+					scale: 2,
+					position: "absolute",
+				}}
+				className={cn(
+					"z-10 inline-block relative text-left text-neutral-900 dark:text-neutral-100 px-2",
+					className,
+				)}
+				key={currentWord}
+			>
+				{currentWord.split("").map((letter, index) => (
+					<motion.span
+						// biome-ignore lint/suspicious/noArrayIndexKey: <explanation>
+						key={currentWord + index}
+						initial={{ opacity: 0, y: 10, filter: "blur(8px)" }}
+						animate={{ opacity: 1, y: 0, filter: "blur(0px)" }}
+						transition={{
+							delay: index * 0.08,
+							duration: 0.4,
+						}}
+						className="inline-block"
+					>
+						{letter}
+					</motion.span>
+				))}
+			</motion.div>
+		</AnimatePresence>
+	);
+};
